moved wait module unit tests to test/util
