+++ test/unit/index.spec.js
@@ -5,5 +5,5 @@ describe('Unit', () => {
   require('./stateView');
   require('./storage');
   require('./sync');
-  require('./util');
+  require('./test/util');
 });
